Split curd tests to crud and extra
